chore: Moved analyses configuration into the analyses module
