+++ ra2ce/analyses/configuration/__init__.py
@@ -1,4 +1,4 @@
-from ra2ce.configuration.analysis.analysis_config_base import (
+from ra2ce.analyses.configuration.analysis_config_base import (
     AnalysisConfigBase,
     AnalysisIniConfigData,
 )
