+++ ra2ce/analyses/configuration/analysis_config_base.py
@@ -23,7 +23,7 @@
 from pathlib import Path
 from typing import Optional
 
-from ra2ce.configuration.analysis.analysis_ini_config_data import AnalysisIniConfigData
+from ra2ce.analyses.configuration.analysis_ini_config_data import AnalysisIniConfigData
 from ra2ce.configuration.config_protocol import ConfigProtocol
