+++ ra2ce/analyses/configuration/readers/analysis_config_reader_factory.py
@@ -24,13 +24,13 @@ from pathlib import Path
 from typing import Optional
 
 from ra2ce.configuration import AnalysisConfigBase
-from ra2ce.configuration.analysis.readers.analysis_config_reader_base import (
+from ra2ce.analyses.configuration.readers.analysis_config_reader_base import (
     AnalysisConfigReaderBase,
 )
-from ra2ce.configuration.analysis.readers.analysis_with_network_config_reader import (
+from ra2ce.analyses.configuration.readers.analysis_with_network_config_reader import (
     AnalysisWithNetworkConfigReader,
 )
-from ra2ce.configuration.analysis.readers.analysis_without_network_config_reader import (
+from ra2ce.analyses.configuration.readers.analysis_without_network_config_reader import (
     AnalysisWithoutNetworkConfigReader,
 )
 from ra2ce.graph.network_config_data.network_config_data import NetworkConfigData
