+++ ra2ce/analyses/configuration/readers/analysis_with_network_config_reader.py
@@ -23,11 +23,11 @@
 from pathlib import Path
 from typing import Optional
 
-from ra2ce.configuration.analysis.analysis_config_base import AnalysisConfigBase
-from ra2ce.configuration.analysis.analysis_ini_config_data import (
+from ra2ce.analyses.configuration.analysis_config_base import AnalysisConfigBase
+from ra2ce.analyses.configuration.analysis_ini_config_data import (
     AnalysisWithNetworkIniConfigData,
 )
-from ra2ce.configuration.analysis.readers.analysis_config_reader_base import (
+from ra2ce.analyses.configuration.readers.analysis_config_reader_base import (
     AnalysisConfigReaderBase,
 )
 from ra2ce.graph.network_config_wrapper import NetworkConfigWrapper
