+++ tests/analyses/configuration/test_analysis_config_factory.py
@@ -1,15 +1,15 @@
 import pytest
 
-from ra2ce.configuration.analysis.analysis_config_factory import AnalysisConfigFactory
-from ra2ce.configuration.analysis.analysis_ini_config_data import (
+from ra2ce.analyses.configuration.analysis_config_factory import AnalysisConfigFactory
+from ra2ce.analyses.configuration.analysis_ini_config_data import (
     AnalysisIniConfigData,
     AnalysisWithNetworkIniConfigData,
     AnalysisWithoutNetworkIniConfigData,
 )
-from ra2ce.configuration.analysis.analysis_with_network_config import (
+from ra2ce.analyses.configuration.analysis_with_network_config import (
     AnalysisWithNetworkConfiguration,
 )
-from ra2ce.configuration.analysis.analysis_without_network_config import (
+from ra2ce.analyses.configuration.analysis_without_network_config import (
     AnalysisWithoutNetworkConfiguration,
 )
 from ra2ce.configuration.ini_config_protocol import IniConfigDataProtocol
