+++ tests/analyses/configuration/test_analysis_with_network_config.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import pytest
 
 from ra2ce.configuration import AnalysisIniConfigData
-from ra2ce.configuration.analysis.analysis_with_network_config import (
+from ra2ce.analyses.configuration.analysis_with_network_config import (
     AnalysisWithNetworkConfiguration,
 )
 from ra2ce.graph.network_config_wrapper import NetworkConfigWrapper
